Minimize and no-std-ify devicemapper-rs-sys.
Run bindgen once (vs. kernel 6.6.62 dm-ioctl.h), copy its output
into devicemapper-rs-sys/src/lib.rs, and remove the bindgen
dependency.

Remove everything to do with System V IPC, and remove the code in
devicemapper-rs proper that used it.  This means the library no longer
attempts to synchronize with udev; this is intentional, as the
minimized devicemapper-rs will be used only in a context where
udev is not available.

Remove all libdevicemapper.h constants except those that are actually
used.  The remaining constants all relate to udev and will be removed
when the remaining vestiges of udev support are removed from the main
library.

Convert all uses of std:: to core::.

Split the tests to a submodule.

Reorganize the no-longer-generated code for human understanding and
copy doc comments out of dm-ioctl.h when available.