Split all tests to separate files.
The convention is now that the tests for src/foo.rs live in
src/tests/foo.rs, but they are still a submodule of crate::foo;
src/foo.rs has

    #[cfg(test)]
    #[path = "tests/foo.rs"]
    mod tests;

The exception is test/support.rs, which is the new name for
testing/mod.rs, now known as the `test_support` submodule of
the crate root.

At the same time, rename src/sys/mod.rs to src/bindings.rs
(and src/sys/test.rs to src/test/bindings.rs).

The only code changes were those necessitated by renaming
crate::testing to crate::test_support and crate::sys to
crate::bindings.

(This is groundwork for a further split, in which the tests that
require root privileges will cease to be unit tests at all.)