Don’t convert io::Error to String eagerly in DmError.
Step one of making DmError less allocation-happy: in the ContextInit
and GeneralIo alternatives, directly wrap the std::io::Error object
rather than converting it to a string eagerly.  This requires us to
stop deriving Clone for DmError, which is fine.

Also rename GeneralIo to RequestConstruction and document that it’s
used for system errors during the construction of request packets.
(I suspect this class of errors cannot actually occur, or can only
occur due to library bugs, but all the marshaling and unmarshaling
logic is slated for overhaul _anyway_ so there’s no point messing
with it now.)