Merge some definitions from bindings.rs to flags.rs and ioctl_cmds.rs.
Specifically, the definitions of the `DM_IOCTL` constant and the
`DmIoctlCmd` enumeration move from bindings.rs to ioctl_cmds.rs,
and the “status bit” constants (`DM_*_FLAG`) and their documentation
are merged into the definition of `DmFlags`.

The numeric values of the DmFlags bits are also changed to match
how they are defined in `dm-ioctl.h` to facilitate future updates
(which will have to be manual), and the missing `DM_IMA_MEASUREMENT`
bit is added to DmFlags.